fix signature mismatch
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@35079 72102866-910b-0410-8b05-ffd578937521

+++ GWorkspace/TShelf/TShelfView.h
@@ -56,7 +56,7 @@
 
 - (void)selectTabItem:(TShelfViewItem *)item;
 
-- (void)selectTabItemAtIndex:(int)index;
+- (void)selectTabItemAtIndex:(NSUInteger)index;
 
 - (void)selectLastItem;
